add some cache utilities
