+++ tests/util/TestUtil.php
@@ -7,7 +7,7 @@ use Plancke\HypixelPHP\HypixelPHP;
 class TestUtil {
 
     static function getHypixelPHP() {
-        $HypixelPHP = new HypixelPHP("");
+        $HypixelPHP = new HypixelPHP('');
         // log to error
         $HypixelPHP->setLogger(new CustomLogger($HypixelPHP));
         // only fetching
